changing from double to BigDecimal

+++ src/main/java/simpleerp/inventory/CreateInventoryDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record CreateInventoryDTO(
 
         @NotNull(message = "Component id is required")
@@ -12,10 +14,10 @@ public record CreateInventoryDTO(
         @NotNull
         @Digits(integer =10, fraction=2, message = "Currently supported (10,2)")
         @Min(value = 0, message = "qty cannot be less than 0")
-        double qtyAvailable,
+        BigDecimal qtyAvailable,
 
         @NotNull
         @Digits(integer =10, fraction=2, message = "Currently supported (10,2)")
         @Min(value = 0, message = "qty cannot be less than 0")
-        double qtyReserved
+        BigDecimal qtyReserved
 ) { }
